Implement dump-screen
This introduces a new `dump-screen` function that dumps the content of the
screen in a ncurses-specific way into a file. The file is controlled
by the config `dump_screen_file` and is rotated each time `dump-screen`
is called.

For now, I have set a maximum of 100 rotations. We can decide to
configure it or we can implement unlimited rotation too.

Implements #3328